requirements.txt

Django==3.2.8
djangorestframework==3.11.2
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.10.1
django-jalali-date==1.0.0

channels==3.0.5
channels_redis==3.4.0
django-redis==5.2.0
redis==3.5.3


drf-nested-routers==0.93.4
drf-yasg==1.17.1
environs==8.0.0
gunicorn==20.1.0
Pillow==8.4.0
pipenv==2021.11.23

validate-email==1.3
whitenoise==6.3.0
